break out module support in TLM PT example, addon
