+++ baseline/pytorch/embeddings.py
@@ -252,6 +252,7 @@ class TransformerLMPooledEmbeddingsModel(TransformerLMEmbeddingsModel):
     def __init__(self, name, **kwargs):
         super().__init__(name=name, **kwargs)
         self.pooling = kwargs.get('pooling', 'cls')
+        reduction_pooling = kwargs.get('reduction_pooling', 'sqrt_length')
         reduction_d_k = kwargs.get('reduction_d_k', self.d_model)
         dropout = kwargs.get('dropout', 0.1)  # use the same dropout as the transformer encoder for reduction layer
         if self.pooling == 'max':
@@ -261,10 +262,10 @@ class TransformerLMPooledEmbeddingsModel(TransformerLMEmbeddingsModel):
         elif self.pooling == 'sqrt_length':
             self.pooling_op = self._sqrt_length_pool
         elif self.pooling == '2HA':
-            self.reduction_layer = TwoHeadConcat(self.d_model, dropout, scale=False, d_k=reduction_d_k)
+            self.reduction_layer = TwoHeadConcat(self.d_model, dropout, scale=False, d_k=reduction_d_k, pooling=reduction_pooling)
             self.pooling_op = self._att_reduction
         elif self.pooling == 'SHA':
-            self.reduction_layer = SingleHeadReduction(self.d_model, dropout, scale=False, d_k=reduction_d_k)
+            self.reduction_layer = SingleHeadReduction(self.d_model, dropout, scale=False, d_k=reduction_d_k, pooling=reduction_pooling)
             self.pooling_op = self._att_reduction
         else:
             self.pooling_op = self._cls_pool
